Extract process termination logic to lifecycle module
